# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SmartKisan AI - Agriculture Advisory API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /crop/calendar:
    get:
      operationId: getCropCalendar
      tags: [crop]
      summary: Get crop calendar and seasonal recommendations
      parameters:
        - name: region
          in: query
          schema:
            type: string
        - name: season
          in: query
          schema:
            type: string
            enum: [kharif, rabi, zaid]
      responses:
        "200":
          description: Crop calendar data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CropCalendar"
  /crop/details:
    get:
      operationId: getCropDetails
      tags: [crop]
      summary: Get detailed information about a specific crop
      parameters:
        - name: crop
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Crop details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CropDetails"
components:
  schemas:
...
    CropCalendarEntry:
      type: object
      properties:
        crop:
          type: string
        season:
          type: string
        sowingMonths:
          type: array
          items:
            type: string
        harvestMonths:
          type: array
          items:
            type: string
        waterRequirement:
          type: string
        soilType:
          type: string
      required:
        - crop
        - season
        - sowingMonths
        - harvestMonths
    CropCalendar:
      type: object
      properties:
        region:
          type: string
        season:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/CropCalendarEntry"
        currentMonth:
          type: string
        recommendations:
          type: array
          items:
            type: string
      required:
        - entries
        - currentMonth
    CropDetails:
      type: object
      properties:
        name:
          type: string
        localNames:
          type: object
          additionalProperties:
            type: string
        description:
          type: string
        season:
          type: string
        duration:
          type: string
        soilRequirement:
          type: string
        waterRequirement:
          type: string
        fertilizers:
          type: array
          items:
            type: string
        diseases:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              symptoms:
                type: string
              treatment:
                type: string
            required:
              - name
              - symptoms
              - treatment
        govtSchemes:
          type: array
          items:
            type: string
        marketPrice:
          type: string
        yieldPerAcre:
          type: string
      required:
        - name
        - description
        - season
        - duration